remove bug with unexisting employee in DB and null reference exception

+++ TestNinja/TestNinja/Mocking/EmployeeStorage.cs
@@ -17,6 +17,7 @@ namespace TestNinja.Mocking
         public void DeleteEmployee(int id)
         {
             var employee = _db.Employees.Find(id);
+            if (employee == null) return;
             _db.Employees.Remove(employee);
             _db.SaveChanges();
         }
